+++ R/WA-estimands.R
@@ -136,7 +136,7 @@ formulaCount <- update.formula(formulaCount,cform)
 dataDmin <- evalTerminal(formulaCount,data=data,time=time,death.code=death.code)
 rrR[,"ratio__"] <- dataDmin[cid$reverseCountid==1,"ratio"]
 if (!is.null(trans)) {
-     rrR[,"ratio__"] <- dataDmin[cid$reverseCountid==1,4]^trans
+     rrR[,"ratio__"] <- rrR[,"ratio__"]^trans
 }
 Yr <- rrR[,"ratio__"]
 
@@ -180,7 +180,7 @@ else {
     treat.modelid <- update.formula(treat.model, . ~ . + cluster(id__))
     treat <- mlogit(treat.modelid, data)
     iidalpha <- lava::iid(treat)
-    pal <- predictmlogit(treat, data, se = 0, response = FALSE)
+    pal <- predict(treat, data, se = 0, response = FALSE)
     ppp <- (pal/pal[, 1])
     spp <- 1/pal[, 1]
 }