+++ R/dreg.R
@@ -39,7 +39,7 @@
 ##' \donttest{ ## Reduce Ex.Timings
 ##' y <- "S*"~"*.width"
 ##' xs <- dreg(dat, y, fun.=phreg)
-##' xs <- dreg(dat, y, fun.=survdiff)
+##' ## xs <- dreg(dat, y, fun.=survdiff)
 ##'
 ##' y <- "S*"~"*.width"
 ##' xs <- dreg(dat, y, x.oneatatime=FALSE, fun.=phreg)
@@ -87,7 +87,7 @@
 ##'
 ##' y <- S1~"*.width"|cluster(id)+factor(species)
 ##' xs <- dreg(dat, y, z.arg="base", fun.=phreg)
-##' xs <- dreg(dat, y, z.arg="base", fun.=coxph)
+##' xs <- dreg(dat, y, z.arg="base", fun.=survival::coxph)
 ##'
 ##' ## under condition with groups
 ##' y <- S1~"*.width"|I(sepal.length>4)